authorization impelemnted

+++ src/api/notes/handler.js
@@ -13,8 +13,9 @@ class NotesHandler {
     async postNoteHandler(request, h) {
         this._validator.validateNotePayload(request.payload)
         const { title = 'untitled', body, tags } = request.payload
+        const { id: credentialId } = request.auth.credentials
 
-        const noteId = await this._service.addNote({ title, body, tags })
+        const noteId = await this._service.addNote({ title, body, tags, owner: credentialId })
 
         const response = h.response({
             status: 'success',
@@ -27,8 +28,9 @@ class NotesHandler {
         return response
     }
 
-    async getNotesHandler() {
-        const notes = await this._service.getNotes()
+    async getNotesHandler(request) {
+        const { id: credentialId } = request.auth.credentials
+        const notes = await this._service.getNotes(credentialId)
 
         return {
             status: 'success',
@@ -38,9 +40,11 @@ class NotesHandler {
         }
     }
 
-    async getNoteByIdHandler(request, h) {
+    async getNoteByIdHandler(request) {
         const { id } = request.params
+        const { id: credentialId } = request.auth.credentials
 
+        await this._service.verifyNoteOwner(id, credentialId)
         const note = await this._service.getNoteById(id)
 
         return {
@@ -51,10 +55,12 @@ class NotesHandler {
         }
     }
 
-    async putNoteByIdHandler(request, h) {
+    async putNoteByIdHandler(request) {
         this._validator.validateNotePayload(request.payload)
         const { id } = request.params
+        const { id: credentialId } = request.auth.credentials
 
+        await this._service.verifyNoteOwner(id, credentialId)
         await this._service.editNoteById(id, request.payload)
 
         return {
@@ -63,9 +69,11 @@ class NotesHandler {
         }
     }
 
-    async deleteNoteByIdHandler(request, h) {
+    async deleteNoteByIdHandler(request) {
         const { id } = request.params
+        const { id: credentialId } = request.auth.credentials
 
+        await this._service.verifyNoteOwner(id, credentialId)
         await this._service.deleteNoteById(id)
 
         return {
